+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # GraphQL Extensions for Apache Jena
 
+# 0.10.2
+
+- Telicent Graph Schema improvements:
+    - Added `nodeFilter` argument to `inRels` and `outRels` fields to filter relationships to only those which relate
+      to specific nodes in the graph
+    - Improved performance of `INCLUDE` filters for `predicateFilter`, and `nodeFilter` by translating these into
+      direct `dsg.stream()` calls for pre-filtering instead of post-filtering a more general `dsg.stream()` call.
+
 # 0.10.1
 
 - Telicent Graph Schema improvements: